+++ Cargo.toml
@@ -27,13 +27,14 @@ development = ["rusty-hook"]
 [workspace]
 members = [
     "crates/primitives",
+    "crates/consensus",
+    "crates/executor",
     "crates/vrrb_core",
     "crates/cli",
     "crates/vrrb_config",
     "crates/vrrb_rpc",
     "crates/block",
     "crates/miner",
-    "crates/state",
     "crates/wallet",
     "crates/mempool",
     "crates/vrrb_lib",
@@ -42,11 +43,14 @@ members = [
     "crates/vrrb_vrf",
     "crates/network",
     "crates/node",
+
     "crates/storage",
+    "crates/storage/lr_trie",
+    "crates/storage/lrdb",
+    "crates/storage/state",
+
     "crates/syncing",
     "crates/quorum",
-    "crates/lrdb",
-    "crates/lr_trie",
     "crates/dkg_engine",
     "crates/telemetry",
     "crates/signer",
@@ -63,23 +67,28 @@ vrrb_config = { path = "crates/vrrb_config" }
 vrrb_rpc = { path = "crates/vrrb_rpc" }
 block = { path = "crates/block" }
 miner = { path = "crates/miner" }
-state = { path = "crates/state" }
 wallet = { path = "crates/wallet" }
 mempool = { path = "crates/mempool" }
+
 # NOTE: potentially deprecated
 vrrb_lib = { path = "crates/vrrb_lib" }
+
 # NOTE: potentially deprecated
 vrrb_cli = { path = "crates/vrrb_cli" }
+
 # NOTE: potentially deprecated
 vrrb_gui = { path = "crates/vrrb_gui" }
 vrrb_vrf = { path = "crates/vrrb_vrf" }
 network = { path = "crates/network" }
 node = { path = "crates/node" }
+
 storage = { path = "crates/storage" }
+state = { path = "crates/storage/state" }
+lrdb = { path = "crates/storage/lrdb" }
+lr_trie = { path = "crates/storage/lr_trie" }
+
 syncing = { path = "crates/syncing" }
 quorum = { path = "crates/quorum" }
-lrdb = { path = "crates/lrdb" }
-lr_trie = { path = "crates/lr_trie" }
 dkg_engine = { path = "crates/dkg_engine" }
 telemetry = { path = "crates/telemetry" }
 signer = { path = "crates/signer" }
@@ -113,6 +122,7 @@ bytebuffer = "0.2.1"
 serde_json = "1.0.64"
 serde = { version = "1.0.144", features = ["derive"] }
 blake3 = "0.3.8"
+log = "0.4.14"
 pickledb = "0.5.1"
 tokio = { version = "1.21.2", features = ["full"] }
 hex = "*"
@@ -135,15 +145,37 @@ jsonrpsee = { version = "0.16.2", features = [
     "ws-client",
 ] }
 bytes = "1.3.0"
-reqwest = "0.11.13"
+reqwest = { version = "0.11.13", features = ["rustls-tls"] }
 sha2 = "0.10.6"
 chrono = "0.4.23"
 lru_time_cache = "0.11.11"
+strum_macros = "0.21.0"
 rocksdb = "0.19.0"
+strum = "0.21.0"
 rayon = "1.6.1"
 rusty-hook = "0.11.2"
 config = "0.13.3"
+serial_test = "0.9.0"
+left-right = "0.11.5"
+keccak-hash = "0.9"
+parking_lot = "0.12"
+rlp = "0.5.1"
+criterion = "0.3.5"
+ethereum-types = "0.13.1"
+
+axum = "0.5.17"
+axum-server = { version = "0.4.4", features = ["tls-rustls"] }
+tower = "0.4.13"
+tower-http = { version = "0.3.4", features = ["trace"] }
+hyper = { version = "0.14.23", features = ["full"] }
+
 
 [dev-dependencies]
 reqwest.workspace = true
 rusty-hook.workspace = true
+serial_test.workspace = true
+rand.workspace = true
+hex.workspace = true
+criterion.workspace = true
+ethereum-types.workspace = true
+uuid.workspace = true
