--- crates/lr_trie/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "lr_trie"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-left-right = "0.11.5"
-patriecia = { git = "https://github.com/vrrb-io/patriecia" }
-serde = { version = "1.0.101", features = ["derive"] }
-keccak-hash = "0.9"
-parking_lot = "0.12"
-rlp = "0.5.1"
-thiserror = "1.0.32"
-bincode = "1.3.3"
-
-[dev-dependencies]
-rand = "0.8.3"
-hex = "0.4.2"
-criterion = "0.3.5"
-ethereum-types = "0.13.1"
-uuid = { version = "0.8.2", features = ["serde", "v4"] }
