--- crates/lrdb/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "lrdb"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-sha2 = "0.10.2"
-serde = { version = "1.0.142", features = ["derive"] }
-thiserror = "1.0.32"
-vrrb_core = { path = "../vrrb_core" }
-lr_trie = { path = "../lr_trie" }
-primitives = { path = "../primitives" }
-patriecia = { git = "https://github.com/vrrb-io/patriecia" }
-bincode = "1.3.3"
-
-[dev-dependencies]
-rand = "0.8.5"
