+++ crates/storage/Cargo.toml
@@ -8,7 +8,31 @@ edition = "2021"
 [dependencies]
 thiserror = { workspace = true }
 primitives = { workspace = true }
-
+ritelinked = { workspace = true }
+sha256 = { workspace = true }
+serde_json = { workspace = true }
+hex = { workspace = true }
+serde = { workspace = true }
+lrdb = { workspace = true }
+mempool = { workspace = true }
+reward = { workspace = true }
+lr_trie = { workspace = true }
+vrrb_core = { workspace = true }
+telemetry = { workspace = true }
+patriecia = { workspace = true }
+rand = { workspace = true }
+derive_builder = { workspace = true }
+rocksdb = { workspace = true }
+left-right = { workspace = true }
+keccak-hash = { workspace = true }
+parking_lot = { workspace = true }
+rlp = { workspace = true }
+bincode = { workspace = true }
 
 [dev-dependencies]
 serial_test = { workspace = true }
+rand = { workspace = true }
+hex = { workspace = true }
+criterion = { workspace = true }
+ethereum-types = { workspace = true }
+uuid = { workspace = true }
