+++ crates/storage/lr_trie/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "lr_trie"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+left-right = { workspace = true }
+patriecia = { workspace = true }
+serde = { workspace = true }
+keccak-hash = { workspace = true }
+parking_lot = { workspace = true }
+rlp = { workspace = true }
+thiserror = { workspace = true }
+bincode = { workspace = true }
+
+[dev-dependencies]
+rand = { workspace = true }
+hex = { workspace = true }
+criterion = { workspace = true }
+ethereum-types = { workspace = true }
+uuid = { workspace = true }
