+++ crates/storage/lr_trie/src/result.rs
