+++ crates/storage/lrdb/src/result.rs
