+++ crates/storage/src/result.rs
@@ -1,0 +1,10 @@
+#[derive(Debug, thiserror::Error)]
+pub enum StorageError {
+    #[error("{0}")]
+    Io(#[from] std::io::Error),
+
+    #[error("{0}")]
+    Other(String),
+}
+
+pub type Result<T> = std::result::Result<T, StorageError>;
